Reorganize files, fix lint and ts errors and add typings
