+++ react/common/layoutContent.ts
@@ -1,41 +1,21 @@
-import CpuUsageLineChart from '../components/Metrics/DataAnalysis/CpuUsageLineChart'
-import MemoryUsageLineChart from '../components/Metrics/DataAnalysis/MemoryUsageLineChart'
+import MemoryUsageLineChart from '../components/Metrics/DataAnalysis/Charts/MemoryUsageLineChart'
 
 export default [
   {
     'ChartType': MemoryUsageLineChart,
     'storedash': {
-      'name': 'runtime',
       'metricParams': {
         'appName': 'render-server',
-        'metricName': 'memoryUsage',
-        'from': 'now-2d/h',
-        'to': 'now-1d/h',
-        'interval': '1h',
-        'operation': 'sum',
         'fields': 'data.summary.external, data.summary.heapTotal, data.summary.heapUsed, data.summary.rss, count',
+        'from': 'now-2d/h',
+        'interval': '1h',
+        'metricName': 'memoryUsage',
+        'operation': 'sum',
         'timezone': '-02:00',
-      }
-    }
+        'to': 'now-1d/h',
+      },
+      'name': 'runtime',
+    },
   },
 ]
 
-
-// {
-//   'ChartType': CpuUsageLineChart,
-//   'storedash': {
-//     'name': 'runtime',
-//     'metricParams': {
-//       'appName': 'render-server',
-//       'metricName': 'cpuUsage',
-//       'from': 'now-2d/h',
-//       'to': 'now-1d/h',
-//       'interval': '1h',
-//       'operation': 'sum',
-//       'fields': 'data.summary.system, data.summary.user, count',
-//     }
-//   }
-// },
-
-// 'appVersion': '8.1.0',
-// 'region': 'aws-us-east-1',
