+++ react/components/Metrics/Controllers/TimeControllers/Relative/index.tsx
@@ -9,5 +9,3 @@ export default class Relative extends Component {
     )
   }
 }
-
-
