+++ react/components/Metrics/DataAnalysis/Charts/MemoryUsageLineChart.tsx
@@ -1,23 +1,23 @@
-import { forEachObjIndexed, includes, map } from 'ramda'
+import { forEachObjIndexed, has, includes, map } from 'ramda'
 import React, { Component, Fragment } from 'react'
 import { Query } from 'react-apollo'
 import { InjectedIntlProps, injectIntl } from 'react-intl'
 import { Spinner } from 'vtex.styleguide'
 
-import dataQuery from '../../../graphql/data.graphql'
+import dataQuery from '../../../../graphql/data.graphql'
 
-import { CHART_PROPERTIES } from '../../../common/constants'
-import { addFormattedTime } from './utils'
-import BlockTitle from './BlockTitle'
-import CustomTooltip from './CustomTooltip'
-import CustomYAxisTick from './CustomYAxisTick'
+import { CHART_PROPERTIES } from '../../../../common/constants'
+import { addFormattedTime } from '../../../../common/utils'
+import BlockTitle from '../CustomElements/BlockTitle'
+import CustomTooltip from '../CustomElements/CustomTooltip'
+import CustomYAxisTick from '../CustomElements/CustomYAxisTick'
+
 
 import {
+  CartesianGrid,
+  Legend,
   Line,
   LineChart,
-  CartesianGrid,
-  Label,
-  Legend,
   ResponsiveContainer,
   Tooltip,
   XAxis,
@@ -30,52 +30,39 @@ interface Props extends InjectedIntlProps {
   metricParams: any
 }
 
+const calculateMean = (value: any, key: string | number | symbol, obj: any) => {
+  const memoryMetrics = ['external', 'heapUsed', 'heapTotal', 'rss']
+  if (includes(key, memoryMetrics)) {
+    obj[key] = Math.round(obj[key] / obj.count)
+  }
+}
 
-const calculateMeanLatency = (chartData: any[]) => {
+const calculateMeanMemory = (chartData: any[]) => {
   return map((chartPoint: any) => {
-    const meanLatency = chartPoint.sum / chartPoint.count
-    return {
-      ...chartPoint,
-      meanLatency,
-    }
+    forEachObjIndexed(calculateMean, chartPoint)
+    return chartPoint
   }, chartData)
 }
 
-// const matchColorByStatusCode = (statusCode: string) => {
-
-// }
-
-// const colors = {
-//   '200': 'Green',
-//   '204': 'Olive',
-//   '301': 'Blue',
-//   '302': 'Navy',
-//   '400': 'Yellow',
-//   '405': 'Orange',
-//   '500': 'Red',
-//   '502': 'Maroon',
-// }
-
-class LatencyLineChart extends Component<Props> {
-  public drawLatencyLine = (chartData: any[]) => {
-
-  }
-
+class MemoryUsageLineChart extends Component<Props> {
   public render = () => {
     const { name, metricParams, intl } = this.props
 
     return (
       <Fragment>
-        <BlockTitle title="Latency (miliseconds) over Time" />
+        <BlockTitle title="Memory (bytes) Consumption over Time" />
 
         <Query query={dataQuery} ssr={false} variables={{ name, params: metricParams }} >
           {({ loading, error, data: { data: rawChartData } }) => {
             let chartData: any
 
             if (!loading) {
-              const stepModifier = metricParams.interval[metricParams.interval.length - 1]
+              let stepModifier = ''
+              if (has('interval', metricParams)) {
+                stepModifier = metricParams.interval[metricParams.interval.length - 1]
+              }
               chartData = addFormattedTime(JSON.parse(rawChartData), intl, stepModifier)
-              chartData = calculateMeanLatency(chartData)
+              chartData = calculateMeanMemory(chartData)
             }
 
             return (
@@ -94,8 +81,11 @@ class LatencyLineChart extends Component<Props> {
                         tick={<CustomYAxisTick />}
                       />
                       <Legend />
-                      {/* <Tooltip content={<CustomTooltip />} />
-                      {this.drawLatencyLine()} */}
+                      <Tooltip content={<CustomTooltip />} />
+                      <Line type="monotone" dataKey="external" stroke="Green" />
+                      <Line type="monotone" dataKey="heapUsed" stroke="Navy" />
+                      <Line type="monotone" dataKey="heapTotal" stroke="Maroon" />
+                      <Line type="monotone" dataKey="rss" stroke="Orange" />
                     </LineChart>
                   </ResponsiveContainer>
                 )
@@ -107,4 +97,4 @@ class LatencyLineChart extends Component<Props> {
   }
 }
 
-export default injectIntl(LatencyLineChart)
+export default injectIntl(MemoryUsageLineChart)
