+++ react/components/Metrics/DataAnalysis/CustomElements/CustomXAxisTick.tsx
@@ -8,7 +8,7 @@ interface Props {
 }
 
 
-const CustomXAxisTick = ({ x, y, payload }: Props) => (
+const CustomXAxisTick = ({ x, y, payload }: Partial<Props>) => (
   <g transform={`translate(${x},${y})`}>
     <text x={0} y={15} dy={0} textAnchor="middle" fill="#666" className="f6">
       {payload.value}
