--- react/components/Metrics/utils.ts
@@ -1,22 +1,0 @@
-export const getAppId = (envControllers: EnvController) => {
-  const {
-    appName,
-    chosenMajor,
-    chosenMinor,
-    chosenPatch,
-  } = envControllers
-  return appName ?
-    `${appName}@${chosenMajor}.${chosenMinor}.${chosenPatch}` :
-    null
-}
-
-export const getAppVersion = (envControllers: EnvController) => {
-  const {
-    chosenMajor,
-    chosenMinor,
-    chosenPatch,
-  } = envControllers
-  return (chosenMajor && chosenMinor && chosenPatch) ?
-    `${chosenMajor}.${chosenMinor}.${chosenPatch}` :
-    null
-}
