--- react/graphql/addSpecToLayout.graphql
@@ -1,12 +1,0 @@
-mutation AddSpecToLayoutWithSpecs($appName: String!, $specLocatorInput: SpecLocatorInput!) {
-  addSpecToLayoutWithSpecs(appName: $appName, $specLocators: $specLocatorInput) {
-    cacheId
-    layoutWithSpecs {
-      spec
-      specLocator {
-        appId
-        specName
-      }
-    }
-  }
-}
