--- react/graphql/deleteSpec.graphql
@@ -1,3 +1,0 @@
-mutation DeleteSpec($specName: String!) {
-  deleteSpec(specName: $specName)
-}
