--- react/graphql/layoutWithSpecs.graphql
@@ -1,14 +1,0 @@
-query LayoutWithSpecs($appName: String!) {
-  layoutWithSpecs (
-    appName:$appName
-  ) {
-    cacheId
-    layoutWithSpecs {
-      spec
-      specLocator {
-        appId
-        specName
-      }
-    }
-  }
-}
