--- react/graphql/saveLayoutWithSpecs.graphql
@@ -1,12 +1,0 @@
-mutation SaveLayoutWithSpecs($appName: String!, $specLocators: [SpecLocatorInput!]!) {
-  saveLayoutWithSpecs(appName:$appName, specLocators: $specLocators) {
-    cacheId
-    layoutWithSpecs {
-      spec
-      specLocator {
-        appId
-        specName
-      }
-    }
-  }
-}
