--- react/graphql/specs.graphql
@@ -1,6 +1,0 @@
-query AvailableSpecs {
-  specs {
-    appId
-    specName
-  }
-}
