--- react/typings/blockTitle.d.ts
@@ -1,2 +1,0 @@
-
-declare module "BlockTitle" {}
