+++ react/typings/events.d.ts
@@ -1,0 +1,8 @@
+export interface EventWithValue extends Event {
+  value: any
+  currentTarget: EventTargetWithValue
+}
+
+export interface EventTargetWithValue extends EventTarget {
+  value: any
+}
