--- react/typings/react-vega.d.ts
@@ -1,1 +1,0 @@
-declare module 'react-vega'
